+++ src/test/resources/com/pestphp/pest/generators/ExpectationGenerator/GeneratedWithMethod.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pest\Expectations;
+namespace Pest;
 
 /**
  * @method bool works()